修复 role 类型统一为 int，优化权限检查 + FIELD_CN 映射
- 所有 role 类型从 short 统一为 int（修复 short 反序列化异常）
- Controller 提取 checkPermission 辅助方法，消除重复权限检查
- FIELD_CN 字段中文映射从 UserEntity 移至 ServiceImpl
- 修复 FIELD_CN 被 MyBatis-Plus 误认为数据库列的问题
- getDetail 支持中文字段名返回

+++ src/main/java/com/example/for_testdemo1/Config/JwtUtil.java
@@ -26,7 +26,7 @@ public class JwtUtil {
     }
 
 
-    public String generateToken(int userId, String account, short role) {
+    public String generateToken(int userId, String account, int role) {
         String token = Jwts.builder()
                 .claim("userId", userId)
                 .claim("account", account)
