+++ src/main/java/com/example/for_testdemo1/Entity/UserEntity.java
@@ -6,13 +6,14 @@ import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 
 import java.time.LocalDateTime;
+import java.util.Map;
 
 @Data
 @TableName("`user`")
 public class UserEntity {
     @TableId(type = IdType.AUTO)
     private int id;
-    private short role;
+    private int role;
     private String account;
     private String password;
     private String name;
@@ -20,4 +21,5 @@ public class UserEntity {
     private String email;
     private LocalDateTime createTime;
     private LocalDateTime updateTime;
+
 }
